Martech_API_Automation/requirements.txt
# Requirements for Python 3.9
ptest
requests
pocoui
airtest
PyHamcrest
PyYaml
kafka-python
